+++ ChangeLog.md
@@ -1,5 +1,6 @@
 # COFantasy: Historique des changements (depuis la version 1.0)
 ## 3.16
+* Support minimal pour le combat de masse.
 * Support pour la dague Croc du Dragon
 
 ## 3.15